# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: trendsAF API Documentation
  description: |
    This API provides CRUD operations for trendsAF, allowing users to interact with the endpoints provided. Each endpoint provides typical HTTP methods such as GET, POST, PUT, and DELETE.
  version: 1.0.0
servers:
  - url: http://143.110.175.227:5080
paths:
  /signup/registration:
    post:
      summary: User Registration
      description: Registers a new user in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                lastname:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Successful registration
  /signup/confirm_email/{token}:
    get:
      summary: Confirm Email
      description: Confirms a user’s email with the provided token.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email confirmed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []